# [Optimization] Add index to billboard placement area