--- a/db/migrate/20230627154435_add_index_to_display_ad_placement_area.rb
+++ b/db/migrate/20230627154435_add_index_to_display_ad_placement_area.rb
@@ -1,0 +1,6 @@
+class AddIndexToDisplayAdPlacementArea < ActiveRecord::Migration[7.0]
+  disable_ddl_transaction!
+  def change
+    add_index :display_ads, :placement_area, algorithm: :concurrently
+  end
+end