<!--
     For Work In Progress Pull Requests, please use the Draft PR feature,
     see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for further details.

     For a timely review/response, please avoid force-pushing additional
     commits if your PR already received reviews or comments.

     Before submitting a Pull Request, please ensure you've done the following:
     - 📖 Read the Forem Contributing Guide: https://developers.forem.com/contributing-guide/forem#create-a-pull-request
     - 📖 Read the Forem Code of Conduct: https://github.com/forem/forem/blob/main/CODE_OF_CONDUCT.md
     - 👷‍♀️ Create small PRs. In most cases this will be possible.
     - ✅ Provide tests for your changes.
     - 📝 Use descriptive commit messages.
     - 📗 Update any related documentation and include any relevant screenshots.

     NOTE: Pull Requests from forked repositories will need to be reviewed by
     a Forem Team member before any CI builds will run. Once your PR is approved
     with a `/ci` reply to the PR, it will be allowed to run subsequent builds without
     manual approval.
-->

## What type of PR is this? (check all applicable)

- [ ] Refactor
- [ ] Feature
- [ ] Bug Fix
- [x] Optimization
- [ ] Documentation Update

## Description

This is a quick optimization with a good likelihood of having a positive impact on performance on a high-traffic endpoint.

As noted below, our average duration on the `/post_sidebar` is currently 72ms. If we shave some time off with this and other common-sense optimizations, we could potentially keep this query under 75ms in the long run, even with some additional logic and scale added in the future.

## QA Instructions, Screenshots, Recordings

https://ui.honeycomb.io/forem.dev/datasets/rails/result/c3wocrvum31

Here is our quintessential billboard unit location `/post_sidebar` performance. We average 72ms per request with a typical P99 of 317ms — not bad. But we make this request hundreds of thousands of times per day and it is worth being proactive around low-hanging fruit.

<img width="881" alt="Screenshot 2023-06-27 at 12 23 03 PM" src="https://github.com/forem/forem/assets/3102842/6a62e0b4-1eb6-4d9b-99df-aa7bf6edb4d5">

We have a snapshot of current performance of past 7 days with 2 hour precision. Any time shaved here is positive, I don't know how much any one index will save, but maybe 5ms?